REF/TST: move non-reindex tests from test_axis_select_reindex.py
