+++ pandas/tests/frame/methods/test_values.py
@@ -1,0 +1,53 @@
+import numpy as np
+
+from pandas import DataFrame, Timestamp, date_range
+import pandas._testing as tm
+
+
+class TestDataFrameValues:
+    def test_values_duplicates(self):
+        df = DataFrame(
+            [[1, 2, "a", "b"], [1, 2, "a", "b"]], columns=["one", "one", "two", "two"]
+        )
+
+        result = df.values
+        expected = np.array([[1, 2, "a", "b"], [1, 2, "a", "b"]], dtype=object)
+
+        tm.assert_numpy_array_equal(result, expected)
+
+    def test_frame_values_with_tz(self):
+        tz = "US/Central"
+        df = DataFrame({"A": date_range("2000", periods=4, tz=tz)})
+        result = df.values
+        expected = np.array(
+            [
+                [Timestamp("2000-01-01", tz=tz)],
+                [Timestamp("2000-01-02", tz=tz)],
+                [Timestamp("2000-01-03", tz=tz)],
+                [Timestamp("2000-01-04", tz=tz)],
+            ]
+        )
+        tm.assert_numpy_array_equal(result, expected)
+
+        # two columns, homogenous
+
+        df["B"] = df["A"]
+        result = df.values
+        expected = np.concatenate([expected, expected], axis=1)
+        tm.assert_numpy_array_equal(result, expected)
+
+        # three columns, heterogeneous
+        est = "US/Eastern"
+        df["C"] = df["A"].dt.tz_convert(est)
+
+        new = np.array(
+            [
+                [Timestamp("2000-01-01T01:00:00", tz=est)],
+                [Timestamp("2000-01-02T01:00:00", tz=est)],
+                [Timestamp("2000-01-03T01:00:00", tz=est)],
+                [Timestamp("2000-01-04T01:00:00", tz=est)],
+            ]
+        )
+        expected = np.concatenate([expected, new], axis=1)
+        result = df.values
+        tm.assert_numpy_array_equal(result, expected)
