+++ pandas/tests/frame/test_axis_select_reindex.py
@@ -4,37 +4,13 @@ import numpy as np
 import pytest
 
 import pandas as pd
-from pandas import Categorical, DataFrame, Index, MultiIndex, Series, date_range, isna
+from pandas import Categorical, DataFrame, Index, Series, date_range, isna
 import pandas._testing as tm
 
 
 class TestDataFrameSelectReindex:
     # These are specific reindex-based tests; other indexing tests should go in
     # test_indexing
-
-    def test_merge_join_different_levels(self):
-        # GH 9455
-
-        # first dataframe
-        df1 = DataFrame(columns=["a", "b"], data=[[1, 11], [0, 22]])
-
-        # second dataframe
-        columns = MultiIndex.from_tuples([("a", ""), ("c", "c1")])
-        df2 = DataFrame(columns=columns, data=[[1, 33], [0, 44]])
-
-        # merge
-        columns = ["a", "b", ("c", "c1")]
-        expected = DataFrame(columns=columns, data=[[1, 11, 33], [0, 22, 44]])
-        with tm.assert_produces_warning(UserWarning):
-            result = pd.merge(df1, df2, on="a")
-        tm.assert_frame_equal(result, expected)
-
-        # join, see discussion in GH 12219
-        columns = ["a", "b", ("a", ""), ("c", "c1")]
-        expected = DataFrame(columns=columns, data=[[1, 11, 0, 44], [0, 22, 1, 33]])
-        with tm.assert_produces_warning(UserWarning):
-            result = df1.join(df2, on="a")
-        tm.assert_frame_equal(result, expected)
 
     def test_reindex(self, float_frame):
         datetime_series = tm.makeTimeSeries(nper=30)
@@ -381,20 +357,6 @@ class TestDataFrameSelectReindex:
         res3 = df.reindex(labels=["b", "a"], axis=0).reindex(labels=["e", "d"], axis=1)
         for res in [res2, res3]:
             tm.assert_frame_equal(res1, res)
-
-    def test_align_int_fill_bug(self):
-        # GH #910
-        X = np.arange(10 * 10, dtype="float64").reshape(10, 10)
-        Y = np.ones((10, 1), dtype=int)
-
-        df1 = DataFrame(X)
-        df1["0.X"] = Y.squeeze()
-
-        df2 = df1.astype(float)
-
-        result = df1 - df1.mean()
-        expected = df2 - df2.mean()
-        tm.assert_frame_equal(result, expected)
 
     def test_reindex_boolean(self):
         frame = DataFrame(
