+++ pandas/tests/frame/test_nonunique_indexes.py
@@ -488,16 +488,6 @@ class TestDataFrameNonuniqueIndexes:
         xp.columns = ["A", "A", "B"]
         tm.assert_frame_equal(rs, xp)
 
-    def test_values_duplicates(self):
-        df = DataFrame(
-            [[1, 2, "a", "b"], [1, 2, "a", "b"]], columns=["one", "one", "two", "two"]
-        )
-
-        result = df.values
-        expected = np.array([[1, 2, "a", "b"], [1, 2, "a", "b"]], dtype=object)
-
-        tm.assert_numpy_array_equal(result, expected)
-
     def test_set_value_by_index(self):
         # See gh-12344
         df = DataFrame(np.arange(9).reshape(3, 3).T)
