+++ pandas/tests/frame/test_timezones.py
@@ -3,7 +3,6 @@ Tests for DataFrame timezone-related methods
 """
 import numpy as np
 import pytest
-import pytz
 
 from pandas.core.dtypes.dtypes import DatetimeTZDtype
 
@@ -14,43 +13,6 @@ from pandas.core.indexes.datetimes import date_range
 
 
 class TestDataFrameTimezones:
-    def test_frame_values_with_tz(self):
-        tz = "US/Central"
-        df = DataFrame({"A": date_range("2000", periods=4, tz=tz)})
-        result = df.values
-        expected = np.array(
-            [
-                [pd.Timestamp("2000-01-01", tz=tz)],
-                [pd.Timestamp("2000-01-02", tz=tz)],
-                [pd.Timestamp("2000-01-03", tz=tz)],
-                [pd.Timestamp("2000-01-04", tz=tz)],
-            ]
-        )
-        tm.assert_numpy_array_equal(result, expected)
-
-        # two columns, homogenous
-
-        df = df.assign(B=df.A)
-        result = df.values
-        expected = np.concatenate([expected, expected], axis=1)
-        tm.assert_numpy_array_equal(result, expected)
-
-        # three columns, heterogeneous
-        est = "US/Eastern"
-        df = df.assign(C=df.A.dt.tz_convert(est))
-
-        new = np.array(
-            [
-                [pd.Timestamp("2000-01-01T01:00:00", tz=est)],
-                [pd.Timestamp("2000-01-02T01:00:00", tz=est)],
-                [pd.Timestamp("2000-01-03T01:00:00", tz=est)],
-                [pd.Timestamp("2000-01-04T01:00:00", tz=est)],
-            ]
-        )
-        expected = np.concatenate([expected, new], axis=1)
-        result = df.values
-        tm.assert_numpy_array_equal(result, expected)
-
     def test_frame_join_tzaware(self):
         test1 = DataFrame(
             np.zeros((6, 3)),
@@ -71,32 +33,6 @@ class TestDataFrameTimezones:
 
         tm.assert_index_equal(result.index, ex_index)
         assert result.index.tz.zone == "US/Central"
-
-    def test_frame_align_aware(self):
-        idx1 = date_range("2001", periods=5, freq="H", tz="US/Eastern")
-        idx2 = date_range("2001", periods=5, freq="2H", tz="US/Eastern")
-        df1 = DataFrame(np.random.randn(len(idx1), 3), idx1)
-        df2 = DataFrame(np.random.randn(len(idx2), 3), idx2)
-        new1, new2 = df1.align(df2)
-        assert df1.index.tz == new1.index.tz
-        assert df2.index.tz == new2.index.tz
-
-        # different timezones convert to UTC
-
-        # frame with frame
-        df1_central = df1.tz_convert("US/Central")
-        new1, new2 = df1.align(df1_central)
-        assert new1.index.tz == pytz.UTC
-        assert new2.index.tz == pytz.UTC
-
-        # frame with Series
-        new1, new2 = df1.align(df1_central[0], axis=0)
-        assert new1.index.tz == pytz.UTC
-        assert new2.index.tz == pytz.UTC
-
-        df1[0].align(df1_central, axis=0)
-        assert new1.index.tz == pytz.UTC
-        assert new2.index.tz == pytz.UTC
 
     @pytest.mark.parametrize("tz", ["US/Eastern", "dateutil/US/Eastern"])
     def test_frame_no_datetime64_dtype(self, tz):
